requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-texttospeech==2.33.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyobjc==12.0
pyobjc-core==12.0
pyobjc-framework-Accessibility==12.0
pyobjc-framework-Accounts==12.0
pyobjc-framework-AddressBook==12.0
pyobjc-framework-AdServices==12.0
pyobjc-framework-AdSupport==12.0
pyobjc-framework-AppleScriptKit==12.0
pyobjc-framework-AppleScriptObjC==12.0
pyobjc-framework-ApplicationServices==12.0
pyobjc-framework-AppTrackingTransparency==12.0
pyobjc-framework-AudioVideoBridging==12.0
pyobjc-framework-AuthenticationServices==12.0
pyobjc-framework-AutomaticAssessmentConfiguration==12.0
pyobjc-framework-Automator==12.0
pyobjc-framework-AVFoundation==12.0
pyobjc-framework-AVKit==12.0
pyobjc-framework-AVRouting==12.0
pyobjc-framework-BackgroundAssets==12.0
pyobjc-framework-BrowserEngineKit==12.0
pyobjc-framework-BusinessChat==12.0
pyobjc-framework-CalendarStore==12.0
pyobjc-framework-CallKit==12.0
pyobjc-framework-Carbon==12.0
pyobjc-framework-CFNetwork==12.0
pyobjc-framework-Cinematic==12.0
pyobjc-framework-ClassKit==12.0
pyobjc-framework-CloudKit==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Collaboration==12.0
pyobjc-framework-ColorSync==12.0
pyobjc-framework-Contacts==12.0
pyobjc-framework-ContactsUI==12.0
pyobjc-framework-CoreAudio==12.0
pyobjc-framework-CoreAudioKit==12.0
pyobjc-framework-CoreBluetooth==12.0
pyobjc-framework-CoreData==12.0
pyobjc-framework-CoreHaptics==12.0
pyobjc-framework-CoreLocation==12.0
pyobjc-framework-CoreMedia==12.0
pyobjc-framework-CoreMediaIO==12.0
pyobjc-framework-CoreMIDI==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-CoreMotion==12.0
pyobjc-framework-CoreServices==12.0
pyobjc-framework-CoreSpotlight==12.0
pyobjc-framework-CoreText==12.0
pyobjc-framework-CoreWLAN==12.0
pyobjc-framework-CryptoTokenKit==12.0
pyobjc-framework-DataDetection==12.0
pyobjc-framework-DeviceCheck==12.0
pyobjc-framework-DeviceDiscoveryExtension==12.0
pyobjc-framework-DictionaryServices==12.0
pyobjc-framework-DiscRecording==12.0
pyobjc-framework-DiscRecordingUI==12.0
pyobjc-framework-DiskArbitration==12.0
pyobjc-framework-DVDPlayback==12.0
pyobjc-framework-EventKit==12.0
pyobjc-framework-ExceptionHandling==12.0
pyobjc-framework-ExecutionPolicy==12.0
pyobjc-framework-ExtensionKit==12.0
pyobjc-framework-ExternalAccessory==12.0
pyobjc-framework-FileProvider==12.0
pyobjc-framework-FileProviderUI==12.0
pyobjc-framework-FinderSync==12.0
pyobjc-framework-FSEvents==12.0
pyobjc-framework-FSKit==12.0
pyobjc-framework-GameCenter==12.0
pyobjc-framework-GameController==12.0
pyobjc-framework-GameKit==12.0
pyobjc-framework-GameplayKit==12.0
pyobjc-framework-HealthKit==12.0
pyobjc-framework-ImageCaptureCore==12.0
pyobjc-framework-InputMethodKit==12.0
pyobjc-framework-InstallerPlugins==12.0
pyobjc-framework-InstantMessage==12.0
pyobjc-framework-Intents==12.0
pyobjc-framework-IntentsUI==12.0
pyobjc-framework-IOBluetooth==12.0
pyobjc-framework-IOBluetoothUI==12.0
pyobjc-framework-IOSurface==12.0
pyobjc-framework-iTunesLibrary==12.0
pyobjc-framework-KernelManagement==12.0
pyobjc-framework-LatentSemanticMapping==12.0
pyobjc-framework-LaunchServices==12.0
pyobjc-framework-libdispatch==12.0
pyobjc-framework-libxpc==12.0
pyobjc-framework-LinkPresentation==12.0
pyobjc-framework-LocalAuthentication==12.0
pyobjc-framework-LocalAuthenticationEmbeddedUI==12.0
pyobjc-framework-MailKit==12.0
pyobjc-framework-MapKit==12.0
pyobjc-framework-MediaAccessibility==12.0
pyobjc-framework-MediaExtension==12.0
pyobjc-framework-MediaLibrary==12.0
pyobjc-framework-MediaPlayer==12.0
pyobjc-framework-MediaToolbox==12.0
pyobjc-framework-Metal==12.0
pyobjc-framework-MetalFX==12.0
pyobjc-framework-MetalKit==12.0
pyobjc-framework-MetalPerformanceShaders==12.0
pyobjc-framework-MetalPerformanceShadersGraph==12.0
pyobjc-framework-MetricKit==12.0
pyobjc-framework-MLCompute==12.0
pyobjc-framework-ModelIO==12.0
pyobjc-framework-MultipeerConnectivity==12.0
pyobjc-framework-NaturalLanguage==12.0
pyobjc-framework-NetFS==12.0
pyobjc-framework-Network==12.0
pyobjc-framework-NetworkExtension==12.0
pyobjc-framework-NotificationCenter==12.0
pyobjc-framework-OpenDirectory==12.0
pyobjc-framework-OSAKit==12.0
pyobjc-framework-OSLog==12.0
pyobjc-framework-PassKit==12.0
pyobjc-framework-PencilKit==12.0
pyobjc-framework-PHASE==12.0
pyobjc-framework-Photos==12.0
pyobjc-framework-PhotosUI==12.0
pyobjc-framework-PreferencePanes==12.0
pyobjc-framework-PushKit==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-QuickLookThumbnailing==12.0
pyobjc-framework-ReplayKit==12.0
pyobjc-framework-SafariServices==12.0
pyobjc-framework-SafetyKit==12.0
pyobjc-framework-SceneKit==12.0
pyobjc-framework-ScreenCaptureKit==12.0
pyobjc-framework-ScreenSaver==12.0
pyobjc-framework-ScreenTime==12.0
pyobjc-framework-ScriptingBridge==12.0
pyobjc-framework-SearchKit==12.0
pyobjc-framework-Security==12.0
pyobjc-framework-SecurityFoundation==12.0
pyobjc-framework-SecurityInterface==12.0
pyobjc-framework-SecurityUI==12.0
pyobjc-framework-SensitiveContentAnalysis==12.0
pyobjc-framework-ServiceManagement==12.0
pyobjc-framework-SharedWithYou==12.0
pyobjc-framework-SharedWithYouCore==12.0
pyobjc-framework-ShazamKit==12.0
pyobjc-framework-Social==12.0
pyobjc-framework-SoundAnalysis==12.0
pyobjc-framework-Speech==12.0
pyobjc-framework-SpriteKit==12.0
pyobjc-framework-StoreKit==12.0
pyobjc-framework-Symbols==12.0
pyobjc-framework-SyncServices==12.0
pyobjc-framework-SystemConfiguration==12.0
pyobjc-framework-SystemExtensions==12.0
pyobjc-framework-ThreadNetwork==12.0
pyobjc-framework-UniformTypeIdentifiers==12.0
pyobjc-framework-UserNotifications==12.0
pyobjc-framework-UserNotificationsUI==12.0
pyobjc-framework-VideoSubscriberAccount==12.0
pyobjc-framework-VideoToolbox==12.0
pyobjc-framework-Virtualization==12.0
pyobjc-framework-Vision==12.0
pyobjc-framework-WebKit==12.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
pytest==9.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-qt==4.5.0
pyttsx3==2.99
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.5.0
